Config changes for SKR Mini E3 v2.0 from u/qwewer1
This is a fork of the Marlin bugfix-2.0.x branch with the changes
from http://reddit.com/u/qwewer1 for the SKR Mini E3 v2.0.  You
should be able to clone this repo and load it into PlatformIO with
VSCode, and then build and use it.

I have tested it with my Ender 3 Pro, purchased in April 2020, and
have had a successful print.

The full guide is at:

  https://www.reddit.com/r/ender3/comments/h8y1ia/marlin_20x_guide_skr_mini_e3_v20_ender_3/